@@ -8,11 +8,11 @@
     "@testing-library/user-event": "^7.1.2",
     "@types/jest": "^24.0.0",
     "@types/node": "^12.0.0",
-    "@types/react": "^16.9.0",
-    "@types/react-dom": "^16.9.0",
+    "@types/react": "^18.0.26",
+    "@types/react-dom": "^18.0.10",
     "@types/react-helmet": "^6.1.0",
     "@types/react-syntax-highlighter": "^11.0.4",
-    "axios": "^0.19.2",
+    "axios": "^1.2.2",
     "clsx": "^1.1.1",
     "date-fns": "^2.16.1",
     "lodash-es": "^4.17.15",
@@ -20,10 +20,10 @@
     "node-sass": "^8.0.0",
     "normalize.css": "^8.0.1",
     "qrcode": "^1.4.4",
-    "react": "17.0.0",
+    "react": "18.2.0",
     "react-cool-onclickoutside": "^1.5.5",
     "react-copy-to-clipboard": "^5.0.2",
-    "react-dom": "17.0.0",
+    "react-dom": "18.2.0",
     "react-helmet": "^6.1.0",
     "react-router-dom": "^5.2.0",
     "react-scripts": "5.0.1",
